requirements.txt
django~=2.1.4
djangorestframework